Make lazy mode errors portable across targets

+++ llama-cpp-bindings/src/model/llama_lazy_mode_parse_error.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Copy, Debug, Eq, PartialEq, thiserror::Error)]
 #[error("unknown llama lazy mode {value}")]
 pub struct LlamaLazyModeParseError {
-    pub value: u32,
+    pub value: i64,
 }
